split out editor logic and encapsulate DOM manipulation better
