+++ src/exampleAssetLoader.js
@@ -1,16 +1,15 @@
-import { createAssetHtml, setSvgDimensions } from './asset.js'
+import { showAsset } from './asset.js'
 import { ironclad, lightbearer, caveLion, berserker } from './exampleAssets.js'
 import { assetScale } from './assetScaling.js'
+import { updateEditorWithAsset } from './editor.js'
 
 //TODO: move assetInput into an editor module and encapsulate better.
 // also consolidate mucking with the assetContainer in one place
 const assetInput = document.querySelector('.interface-input')
-const assetContainer = document.querySelector(".assets")
 
 const showSingleAssetExample = (asset) => {
-    assetInput.value = JSON.stringify(asset, null, 2)
-    assetContainer.innerHTML = createAssetHtml(asset, assetScale)
-    setSvgDimensions()
+    updateEditorWithAsset(asset)
+    showAsset(asset, assetScale)
 }
 
 export const loadDefaultExampleAsset = () => showSingleAssetExample(ironclad)
